requirements.txt
couchdb
Mastodon.py